move the interface and factory to same file
