+++ ImageHosting.Storage/Features/Images/Services/MetadataUploadCommand.cs
@@ -6,6 +6,19 @@ using ImageHosting.Storage.Generic;
 
 namespace ImageHosting.Storage.Features.Images.Services;
 
+public interface IMetadataUploadCommandFactory
+{
+    IRollbackCommand CreateCommand(Guid userId, Guid imageId, string objectName, bool hidden, DateTime uploadedAt);
+}
+
+public class MetadataUploadCommandFactory(IMetadataService metadataService) : IMetadataUploadCommandFactory
+{
+    public IRollbackCommand CreateCommand(Guid imageId, Guid userId, string objectName, bool hidden, DateTime uploadedAt)
+    {
+        return new MetadataUploadCommand(metadataService, imageId, userId, objectName, hidden, uploadedAt);
+    }
+}
+
 public class MetadataUploadCommand(
     IMetadataService metadataService,
     Guid userId,
